moved shooter speed logic to commandfactory, added comments to gravity compensation method
