+++ src/main/java/frc/robot/CommandFactory.java
@@ -8,4 +8,18 @@ public class CommandFactory {
     public CommandFactory(Arm arm) {
         this.arm = arm;
     }
+
+    public double getSpeedFromArmHeight() {
+        double speedFromArmHeight = 0;
+        if (arm.getEncoderPosition() <= 0.37) {
+            speedFromArmHeight = Constants.ShooterConstants.SUBWOOFER_SPEED;
+        } else if (arm.getEncoderPosition() > 0.37 & arm.getEncoderPosition() <= 0.76) {
+            speedFromArmHeight = Constants.ShooterConstants.SPEAKER_SPEED;
+        } else if (arm.getEncoderPosition() > 0.76 & arm.getEncoderPosition() <= 1) {
+            speedFromArmHeight = Constants.ShooterConstants.FAR_AWAY_SPEED;
+        } else if (arm.getEncoderPosition() > 1) {
+            speedFromArmHeight = Constants.ShooterConstants.AMP_SPEED;
+        }
+        return speedFromArmHeight;
+    }
 }
