+++ src/main/java/frc/robot/subsystems/climber/ClimberReal.java
@@ -1,6 +1,5 @@
 package frc.robot.subsystems.climber;
 
-import com.revrobotics.CANSparkBase;
 import com.revrobotics.CANSparkLowLevel.MotorType;
 import com.revrobotics.CANSparkMax;
 import com.revrobotics.REVLibError;
